requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.18.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
